# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A custom PaaS platform for automated application deployment.\nDeploy applications from Git repositories to Kubernetes with zero-configuration.",
        "title": "App Deployer API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/alvesdmateus/app-deployer"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/analyze": {
            "post": {
                "description": "Analyze source code at a given path to detect language, framework, and dependencies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analyzer"
                ],
                "summary": "Analyze source code",
                "parameters": [
                    {
                        "description": "Path to analyze",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AnalyzeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/analyzer.AnalysisResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/analyze/upload": {
            "post": {
                "description": "Upload source code as a file and analyze it to detect language, framework, and dependencies",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analyzer"
                ],
                "summary": "Upload and analyze source code",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Source code archive (zip, tar.gz)",
                        "name": "source",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/analyzer.AnalysisResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/builds": {
            "post": {
                "description": "Build a container image from source code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "builder"
                ],
                "summary": "Build container image",
                "parameters": [
                    {
                        "description": "Build request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BuildImageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.BuildImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/builds/generate-dockerfile": {
            "post": {
                "description": "Analyze source code and generate an optimized Dockerfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "builder"
                ],
                "summary": "Generate Dockerfile",
                "parameters": [
                    {
                        "description": "Source path",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.GenerateDockerfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GenerateDockerfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/deployments": {
            "get": {
                "description": "Retrieve a paginated list of all deployments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "List all deployments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of results (default: 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ListDeploymentsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new deployment configuration. Optionally triggers immediate provisioning if image_tag is provided.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Create a new deployment",
                "parameters": [
                    {
                        "description": "Deployment configuration",
                        "name": "deployment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateDeploymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.DeploymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/deployments/status/{status}": {
            "get": {
                "description": "Retrieve all deployments with a specific status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Get deployments by status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment status (PENDING, QUEUED, BUILDING, PROVISIONING, DEPLOYING, HEALTHY, FAILED, etc.)",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.DeploymentResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/deployments/{id}": {
            "get": {
                "description": "Retrieve detailed information about a specific deployment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Get a deployment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DeploymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a deployment and its associated infrastructure. If infrastructure exists, triggers async destruction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Delete a deployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deployment deleted (no infrastructure)",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "202": {
                        "description": "Destruction initiated (has infrastructure)",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/orchestrator/stats": {
            "get": {
                "description": "Retrieve statistics about the job queues (provision, deploy, destroy, rollback)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator"
                ],
                "summary": "Get queue statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.QueueStatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Orchestration service unavailable",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "analyzer.AnalysisResult": {
            "type": "object",
            "properties": {
                "build_command": {
                    "type": "string"
                },
                "build_tool": {
                    "$ref": "#/definitions/analyzer.BuildTool"
                },
                "confidence": {
                    "type": "number"
                },
                "dependencies": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "dev_dependencies": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "framework": {
                    "$ref": "#/definitions/analyzer.Framework"
                },
                "has_dockerfile": {
                    "type": "boolean"
                },
                "language": {
                    "$ref": "#/definitions/analyzer.Language"
                },
                "port": {
                    "type": "integer"
                },
                "runtime": {
                    "type": "string"
                },
                "start_command": {
                    "type": "string"
                }
            }
        },
        "analyzer.BuildTool": {
            "type": "string",
            "enum": [
                "npm",
                "yarn",
                "pnpm",
                "go",
                "pip",
                "poetry",
                "maven",
                "gradle",
                "cargo",
                "unknown"
            ],
            "x-enum-varnames": [
                "BuildToolNPM",
                "BuildToolYarn",
                "BuildToolPNPM",
                "BuildToolGo",
                "BuildToolPip",
                "BuildToolPoetry",
                "BuildToolMaven",
                "BuildToolGradle",
                "BuildToolCargo",
                "BuildToolUnknown"
            ]
        },
        "analyzer.Framework": {
            "type": "string",
            "enum": [
                "gin",
                "echo",
                "chi",
                "fiber",
                "express",
                "nestjs",
                "nextjs",
                "koa",
                "fastify",
                "flask",
                "django",
                "fastapi",
                "springboot",
                "quarkus",
                "unknown"
            ],
            "x-enum-varnames": [
                "FrameworkGin",
                "FrameworkEcho",
                "FrameworkChi",
                "FrameworkFiber",
                "FrameworkExpress",
                "FrameworkNestJS",
                "FrameworkNextJS",
                "FrameworkKoa",
                "FrameworkFastify",
                "FrameworkFlask",
                "FrameworkDjango",
                "FrameworkFastAPI",
                "FrameworkSpringBoot",
                "FrameworkQuarkus",
                "FrameworkUnknown"
            ]
        },
        "analyzer.Language": {
            "type": "string",
            "enum": [
                "go",
                "nodejs",
                "python",
                "java",
                "rust",
                "ruby",
                "php",
                "dotnet",
                "unknown"
            ],
            "x-enum-varnames": [
                "LanguageGo",
                "LanguageNodeJS",
                "LanguagePython",
                "LanguageJava",
                "LanguageRust",
                "LanguageRuby",
                "LanguagePHP",
                "LanguageDotNet",
                "LanguageUnknown"
            ]
        },
        "api.AnalyzeRequest": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "api.BuildImageRequest": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "deployment_id": {
                    "type": "string"
                },
                "source_path": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "api.BuildImageResponse": {
            "type": "object",
            "properties": {
                "build_id": {
                    "type": "string"
                },
                "build_log": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "image_digest": {
                    "type": "string"
                },
                "image_tag": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "api.CreateDeploymentRequest": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "cloud": {
                    "type": "string"
                },
                "image_tag": {
                    "description": "Optional: if provided, triggers immediate provisioning",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "description": "Optional: defaults to 8080",
                    "type": "integer"
                },
                "region": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "api.DeploymentResponse": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "cloud": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deployed_at": {
                    "type": "string"
                },
                "external_ip": {
                    "type": "string"
                },
                "external_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.GenerateDockerfileRequest": {
            "type": "object",
            "properties": {
                "source_path": {
                    "type": "string"
                }
            }
        },
        "api.GenerateDockerfileResponse": {
            "type": "object",
            "properties": {
                "analysis": {
                    "$ref": "#/definitions/analyzer.AnalysisResult"
                },
                "dockerfile": {
                    "type": "string"
                },
                "framework": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                }
            }
        },
...
        "api.ListDeploymentsResponse": {
            "type": "object",
            "properties": {
                "deployments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.DeploymentResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.QueueStatsResponse": {
            "type": "object",
            "properties": {
                "deploy": {
                    "type": "integer"
                },
                "destroy": {
                    "type": "integer"
                },
                "provision": {
                    "type": "integer"
                },
                "rollback": {
                    "type": "integer"
                }
            }
        },
...
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}